[Update]modify shared memory semaphore flag.
1.修改共享内存信号量的flag。

+++ camera_service/include/camera_service/shared_memory_op.hpp
@@ -25,7 +25,7 @@
 
 struct SharedImage
 {
-  double img_stamp;
+  uint64_t img_stamp;
   unsigned char img_data[IMAGE_SIZE];
 };
